third_party/WebKit/LayoutTests/platform/win/inspector/curl-command-expected.txt

Tests curl command generation

curl "http://example.org/path" --compressed
curl "http://example.org/path" --data-binary "123" --compressed
curl "http://example.org/path" -H "Content-Type: application/x-www-form-urlencoded" --data "1&b" --compressed
curl "http://example.org/path" -H "Content-Type: application/json" --data-binary "{""a"":1}" --compressed
curl "http://example.org/path" -H "Content-Type: application/binary" --data-binary "1234"^
"00'""!" --compressed
curl "http://example.org/path" -H "Content-Type: application/binary" --data-binary "1234"^
"00'""!" --compressed
curl "http://example.org/path" -H "Content-Type: application/binary" --data-binary "ÿ	܀" --compressed
curl "http://labs.ft.com/?querystring=\[\]\{\}" --compressed
curl "http://example.org/path" -H "Content-Type: application/binary" --data-binary ""%"PATH"%"$PATH" --compressed
